+++ CHANGES.rst
@@ -7,6 +7,8 @@ Unreleased
 
 - Added ``croud projects edit`` command to change a specified project name. At the moment, only the name of the project may be changed
 
+- Removed the ``croud monitoring`` command and all its subcommands.
+
 0.21.0 - 2019/12/20
 ===================